Enable reference to foreign key in other app
Really grateful for this project, made it really easy for me to import bulk data. Trying to import data into my HR app (foreign key reference to Auth.User) and it kept giving me an error. Made a minor edit to fix the problem. The fix is not perfect; it assumes model names are distinct.