Correct logic in save_csvimport
This is a speculative fix, since I'm unsure exactly what is expected to happen.
If |instance| is None, we create a new CSVImport, move the properties to |instance|,
and then save it. When used as a management command, however, |instance| is not None,
so we run into an exception when we try to save it.

It seems like prior testing never hit this code path, so reverse the logic and
bail out if |instance| is set.